Refactor: Extract `detect::linux` in crate `detect-targets`
